finished chapter 3. Started chapter 4

+++ 1-Learning_C++/3_07_challenge/records.h
@@ -23,7 +23,7 @@ public:
 	Course(int the_id, std::string the_name, unsigned char the_credits);
 	int get_id();
 	std::string get_name();
-	int get_credits();
+	uint8_t get_credits();
 };
 
 class Grade{
